remove styled, split into editor
